// Frontend/src/api/models/MenuItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ServePOS.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MenuItem
 */
export interface MenuItem {
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    id: number;
    /**
     * 
     * @type {MenuItemTypeEnum}
     * @memberof MenuItem
     */
    type: MenuItemTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    name: string | null;
    /**
     * 
     * @type {string}
     * @memberof MenuItem
     */
    shortName: string | null;
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    availableQuantity: number | null;
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    soldQuantity: number;
    /**
     * 
     * @type {number}
     * @memberof MenuItem
     */
    sortIndex: number;
}


/**
 * @export
 */
export const MenuItemTypeEnum = {
    Drink: 'Drink',
    Food: 'Food',
    Dessert: 'Dessert',
    Special: 'Special'
} as const;
export type MenuItemTypeEnum = typeof MenuItemTypeEnum[keyof typeof MenuItemTypeEnum];


/**
 * Check if a given object implements the MenuItem interface.
 */
export function instanceOfMenuItem(value: object): value is MenuItem {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('shortName' in value) || value['shortName'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    if (!('availableQuantity' in value) || value['availableQuantity'] === undefined) return false;
    if (!('soldQuantity' in value) || value['soldQuantity'] === undefined) return false;
    if (!('sortIndex' in value) || value['sortIndex'] === undefined) return false;
    return true;
}

export function MenuItemFromJSON(json: any): MenuItem {
    return MenuItemFromJSONTyped(json, false);
}

export function MenuItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuItem {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'type': json['type'],
        'name': json['name'],
        'shortName': json['shortName'],
        'price': json['price'],
        'availableQuantity': json['availableQuantity'],
        'soldQuantity': json['soldQuantity'],
        'sortIndex': json['sortIndex'],
    };
}

export function MenuItemToJSON(json: any): MenuItem {
    return MenuItemToJSONTyped(json, false);
}

export function MenuItemToJSONTyped(value?: MenuItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'name': value['name'],
        'shortName': value['shortName'],
        'price': value['price'],
        'availableQuantity': value['availableQuantity'],
        'soldQuantity': value['soldQuantity'],
        'sortIndex': value['sortIndex'],
    };
}